Fix auth and error pages
