+++ frontend/src/pages/SignUpPage.tsx
@@ -5,17 +5,17 @@ import logo from '../assets/img/logo/logo-trans-bg.png'
 import { FaUserAlt, FaExclamationCircle, FaLock } from 'react-icons/fa'
 import { useSignUpMutation } from '../redux/services/authentication'
 import { getStoredJwtToken } from '../redux/slices/auth'
-import Layout from '@layouts'
+import Layout from '@layouts/index'
 
 function SignUpPage() {
 	const navigate = useNavigate()
-	const [error, setError] = useState()
+	const [error, setError] = useState<string | undefined>()
 	const [username, setUsername] = useState('')
 	const [password, setPassword] = useState('')
 	const [confirm, setConfirm] = useState('')
 	const [fetchLoginApi] = useSignUpMutation()
 
-	const auth = useSelector((state) => state.auth)
+	const auth = useSelector((state: any) => state.auth)
 	const dispatch = useDispatch()
 
 	useEffect(() => {
@@ -29,7 +29,7 @@ function SignUpPage() {
 		}
 	}, [auth.isLoggedIn])
 
-	const handleSignUp = async (e) => {
+	const handleSignUp = async (e: any) => {
 		e.preventDefault()
 		setError(undefined)
 
@@ -37,7 +37,9 @@ function SignUpPage() {
 			const signUpResult = await fetchLoginApi({ username, password, confirm })
 
 			if ('error' in signUpResult) {
-				return setError(signUpResult?.error?.data.message)
+				//TODO: Fix this
+				// return setError(signUpResult?.error?.data.message)
+				return setError('')
 			}
 
 			// return navigate("/login")
